requirements.txt
pymongo==4.3.3
python-dotenv==0.21.0
QueryStateLib==1.1.4
web3==5.31.1
